+++ src/Features/Core/Portable/TaskList/AbstractTaskListService.cs
@@ -10,10 +10,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Text;
+using Microsoft.CodeAnalysis.TodoComments;
 
-namespace Microsoft.CodeAnalysis.TodoComments
+namespace Microsoft.CodeAnalysis.TaskList
 {
-    internal abstract class AbstractTodoCommentService : ITodoCommentService
+    internal abstract class AbstractTaskListService : ITodoCommentService
     {
         protected abstract bool PreprocessorHasComment(SyntaxTrivia trivia);
         protected abstract bool IsSingleLineComment(SyntaxTrivia trivia);
@@ -63,9 +64,7 @@ namespace Microsoft.CodeAnalysis.TodoComments
         {
             var index = GetCommentStartingIndex(message);
             if (index >= message.Length)
-            {
                 return;
-            }
 
             var normalized = GetNormalizedText(message);
             foreach (var commentDescriptor in commentDescriptors)
@@ -78,14 +77,12 @@ namespace Microsoft.CodeAnalysis.TodoComments
                     continue;
                 }
 
-                if ((message.Length > index + token.Length) && IsIdentifierCharacter(message[index + token.Length]))
-                {
+                if (message.Length > index + token.Length && IsIdentifierCharacter(message[index + token.Length]))
                     // they wrote something like:
                     // todoboo
                     // instead of
                     // todo
                     continue;
-                }
 
                 todoList.Add(new TodoComment(commentDescriptor, message[index..], start + index));
             }
@@ -128,9 +125,7 @@ namespace Microsoft.CodeAnalysis.TodoComments
 
             var length = fullSpan.End - endLine.Start;
             if (length >= postfixLength)
-            {
                 length -= postfixLength;
-            }
 
             var endMessage = text.ToString(new TextSpan(endLine.Start, length));
             AppendTodoCommentInfoFromSingleLine(commentDescriptors, endMessage, endLine.Start, todoList);
